requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
Pillow>=10.0.0
google-generativeai>=0.3.0
